feat: update token transfer table, include crc20, crc721, crc1155
