+++ src/app/components/TablePanel/Datepicker.tsx
@@ -242,6 +242,7 @@ export const TableSearchDatepicker = ({
             url: location.pathname,
             query: {
               ...queries,
+              skip: '0',
               minTimestamp,
               maxTimestamp,
             },
