+++ src/app/containers/TokenDetail/Transfers.tsx
@@ -1,31 +1,19 @@
 import React from 'react';
 import styled from 'styled-components/macro';
 import { useTranslation } from 'react-i18next';
-import { media, useBreakpoint } from 'styles/media';
 import { translations } from 'locales/i18n';
-import { useHistory, useLocation } from 'react-router-dom';
-import queryString from 'query-string';
-import { isAddress, isHash, toThousands } from 'utils';
 import { TabsTablePanel } from 'app/components/TabsTablePanel/Loadable';
-import {
-  TableSearchDatepicker,
-  TableSearchInput,
-} from 'app/components/TablePanel';
-import { tokenColunms } from 'utils/tableColumns';
 import { cfxTokenTypes } from 'utils/constants';
-import { useAge } from 'utils/hooks/useAge';
 import { Card } from 'app/components/Card';
 import { LineChart as Chart } from 'app/components/Chart/Loadable';
-import { DownloadCSV } from 'app/components/DownloadCSV/Loadable';
-import { useMessages } from '@cfxjs/react-ui';
-import _ from 'lodash';
 import {
   ContractContent,
   CheckCircleIcon,
 } from '../AddressContractDetail/ContractContent';
-import { useContract } from '../../../utils/api';
+import { useContract } from 'utils/api';
 import AlertCircle from '@zeit-ui/react-icons/alertCircle';
 
+import { Transfers as TokenTransfers } from 'app/containers/Tokens/Loadable';
 import { Holders } from './Holders';
 
 interface TransferProps {
@@ -46,7 +34,6 @@ interface Query {
 }
 
 export function Transfers({ tokenData }: { tokenData: TransferProps }) {
-  const [, setMessage] = useMessages();
   const {
     tokenName,
     address: tokenAddress,
@@ -58,11 +45,7 @@ export function Transfers({ tokenData }: { tokenData: TransferProps }) {
       : '',
     isRegistered,
   } = tokenData;
-  const bp = useBreakpoint();
   const { t } = useTranslation();
-  const location = useLocation();
-  const history = useHistory();
-  const [ageFormat, toggleAgeFormat] = useAge();
 
   const { data: contractInfo } = useContract(tokenAddress, [
     'name',
@@ -84,149 +67,17 @@ export function Transfers({ tokenData }: { tokenData: TransferProps }) {
     'verifyInfo',
   ]);
 
-  let {
-    page = 1,
-    pageSize = 10,
-    accountAddress: filterAddr,
-    transactionHash: filterHash,
-    tokenId: filterTokenId,
-    tab: currentTab,
-    ...others
-  } = queryString.parse(location.search);
-
-  const filter =
-    (filterAddr as string) ||
-    (filterHash as string) ||
-    (filterTokenId as string) ||
-    '';
-
-  const onFilter = (filter: string) => {
-    let object: Query = {};
-
-    if (filter) {
-      if (isAddress(filter)) {
-        object.accountAddress = filter;
-      } else if (isHash(filter)) {
-        object.transactionHash = filter;
-      } else if (transferType !== cfxTokenTypes.erc20 && _.isInteger(+filter)) {
-        object.tokenId = filter;
-      } else {
-        setMessage({
-          text: t(translations.token.transferList.searchError),
-        });
-        return;
-      }
-    }
-
-    const urlWithQuery = queryString.stringifyUrl({
-      url: location.pathname,
-      query: {
-        ...others,
-        page: page as string,
-        pageSize: pageSize as string,
-        ...object,
-      },
-    });
-    history.push(urlWithQuery);
-  };
-
-  let columnsWidth = [3, 6, 6, 4, 4];
-  let columns = [
-    tokenColunms.txnHash,
-    tokenColunms.from,
-    tokenColunms.to,
-    {
-      ...tokenColunms.quantity,
-      render: (value, row, index) =>
-        tokenColunms.quantity.render(value, row, index, {
-          decimals,
-        }),
-    },
-    tokenColunms.age(ageFormat, toggleAgeFormat),
-  ].map((item, i) => ({ ...item, width: columnsWidth[i] }));
-
-  if (transferType === cfxTokenTypes.erc721) {
-    columnsWidth = [3, 6, 6, 4, 3];
-    columns = [
-      tokenColunms.txnHash,
-      tokenColunms.from,
-      tokenColunms.to,
-      tokenColunms.tokenId(),
-      tokenColunms.age(ageFormat, toggleAgeFormat),
-    ].map((item, i) => ({ ...item, width: columnsWidth[i] }));
-  }
-  if (transferType === cfxTokenTypes.erc1155) {
-    columnsWidth = [3, 7, 7, 3, 4, 4];
-    columns = [
-      tokenColunms.txnHash,
-      tokenColunms.from,
-      tokenColunms.to,
-      tokenColunms.quantity,
-      tokenColunms.tokenId(tokenAddress),
-      tokenColunms.age(ageFormat, toggleAgeFormat),
-    ].map((item, i) => ({ ...item, width: columnsWidth[i] }));
-  }
-
   const tabs: any = [
     {
       value: 'transfers',
       action: 'tokenTransfers',
       label: t(translations.token.transfers),
-      // address filter contract transfers events
-      // accountAddress filter from or to transfers, regard accountAddress as ordinary address
-      url: `/transfer?address=${tokenAddress}&transferType=${transferType}`,
-      table: {
-        columns: columns,
-        rowKey: (row, index) => `${row.transactionHash}${index}`,
-      },
-      tableHeader: ({ total }) => (
-        <StyledSearchAreaWrapper>
-          <StyledTotalWrapper>
-            {t(
-              total > 10000
-                ? translations.general.totalRecordLimit
-                : translations.general.totalRecord,
-              {
-                total: toThousands(total),
-              },
-            )}
-          </StyledTotalWrapper>
-          <div className="token-search-container">
-            <TableSearchInput
-              onFilter={onFilter}
-              filter={filter}
-              placeholder={
-                transferType === cfxTokenTypes.erc20
-                  ? `${t(
-                      translations.general.searchInputPlaceholder.txnHash,
-                    )} / ${t(
-                      translations.general.searchInputPlaceholder.holderAddress,
-                    )}`.replace(bp === 's' ? / \/ /gi : '/', '/')
-                  : `${t(
-                      translations.general.searchInputPlaceholder.txnHash,
-                    )} / ${t(
-                      translations.general.searchInputPlaceholder.holderAddress,
-                    )} / ${t(
-                      translations.general.searchInputPlaceholder.tokenID,
-                    )}`.replace(bp === 's' ? / \/ /gi : '/', '/')
-              }
-            />
-            <TableSearchDatepicker />
-          </div>
-        </StyledSearchAreaWrapper>
-      ),
-      tableFooter: (
-        <DownloadCSV
-          url={queryString.stringifyUrl({
-            url: '/v1/report/transfer',
-            query: {
-              transferType: transferType,
-              address: tokenAddress,
-              limit: '5000',
-              reverse: 'true',
-            },
-          })}
-        />
+      content: (
+        <TokenTransfers
+          type={transferType}
+          address={tokenAddress}
+          decimals={decimals}
+        ></TokenTransfers>
       ),
     },
   ];
@@ -243,11 +94,11 @@ export function Transfers({ tokenData }: { tokenData: TransferProps }) {
       label: t(translations.token.holders),
       content: (
         <Holders
-          url={`/stat/tokens/holder-rank?address=${tokenAddress}&reverse=true&orderBy=balance`}
           type={transferType}
           decimals={decimals}
           price={price}
           totalSupply={totalSupply}
+          address={tokenAddress}
         />
       ),
     });
@@ -308,28 +159,6 @@ export function Transfers({ tokenData }: { tokenData: TransferProps }) {
   return transferType ? <TabsTablePanel tabs={tabs} /> : null;
 }
 
-const StyledSearchAreaWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex-wrap: wrap;
-
-  .token-search-container {
-    flex-grow: 1;
-    display: flex;
-    justify-content: flex-end;
-
-    ${media.s} {
-      justify-content: flex-start;
-      margin-top: 0.3571rem;
-    }
-  }
-
-  ${media.s} {
-    justify-content: flex-start;
-  }
-`;
-
 const StyledTabWrapper = styled.div`
   .card {
     padding: 0.3571rem !important;
@@ -340,12 +169,5 @@ const StyledTabWrapper = styled.div`
         box-shadow: none !important;
       }
     }
-  }
-`;
-
-const StyledTotalWrapper = styled.div`
-  padding: 8px;
-  ${media.s} {
-    width: 100%;
   }
 `;
