+++ src/app/containers/Tokens/Loadable.tsx
@@ -8,3 +8,8 @@ export const Tokens = lazyLoad(
   () => import('./index'),
   module => module.Tokens,
 );
+
+export const Transfers = lazyLoad(
+  () => import('./Transfers'),
+  module => module.Transfers,
+);
