+++ src/locales/zh_cn/translation.json
@@ -505,6 +505,7 @@
       "epoch": "纪元"
     },
     "totalRecord": "共计 {{total}} 个记录",
+    "totalRecordWithLimit": "仅展示最新的 {{limit}} 条记录（共计 {{total}} 条记录）",
     "totalHolders": "共计 {{total}} 个持有人",
     "totalRecordLimit": "仅展示最新的 10k 条记录（共计 {{total}} 条记录）",
     "remark": "说明：",
